test: add RetryConfig tests for default values, parameter reading, and backoff calculations
