+++ nodejs/test/bulkPulling/dsn.test.ts
@@ -454,9 +454,7 @@ describe("dsn", () => {
             const dsn = parse(
                 "ws://root:taosdata@localhost:6041/mydb?token=t1&bearer_token=b1&td.connect.token=c1&timezone=UTC"
             );
-
             const masked = JSON.parse(dsn.toString());
-
             expect(masked.password).toBe("[REDACTED]");
             expect(masked.params.token).toBe("[REDACTED]");
             expect(masked.params.bearer_token).toBe("[REDACTED]");
@@ -468,10 +466,7 @@ describe("dsn", () => {
             const dsn = parse(
                 "ws://root:taosdata@localhost:6041?Token=t1&Bearer_Token=b1&TD.CONNECT.TOKEN=c1"
             );
-
             const masked = JSON.parse(dsn.toString());
-            console.log(masked);
-
             expect(masked.params.Token).toBe("[REDACTED]");
             expect(masked.params.Bearer_Token).toBe("[REDACTED]");
             expect(masked.params["TD.CONNECT.TOKEN"]).toBe("[REDACTED]");
